refactor: rename components to webview and move WebView into it
