+++ src/lib.rs
@@ -29,5 +29,3 @@ pub use errors::{Error, Result};
 pub use verso::Verso;
 /// Re-exporting Winit for the sake of convenience.
 pub use winit;
-/// Context
-pub mod components;
