+++ src/webview/mod.rs
@@ -1,3 +1,6 @@
+mod webview;
+/// WebView
+pub use webview::{Panel, WebView};
 /// Context Menu
 pub mod context_menu;
 /// Prompt Dialog
